Feat/goal: 미션 & 목표 관련 API

+++ src/main/java/com/luckit/goal/domain/GoalRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GoalRepository extends JpaRepository<Goal, Long> {
+public interface GoalRepository extends JpaRepository<Goal, Integer> {
 
     List<Goal> findAllByUser(User user);
